feature: Workaround to allow passing an ARN for a Secret in addition to the name
ARNs contain a ':' characters, which the upstream parsing logic in Kafka already treats as a separator and will not handle correctly.

This allows you to mask these characters with an alternative (say '|'), which is safely passed through and then converted back before retrieving the Secret.

A new configuration param is exposed (`separator.replacement`) to allow you to specify the replacement character. This will simply be search and replaced later (if the replacement character does not exist in the string, no changes will be made or error thrown).

Fixes aws-samples/msk-config-providers#22